fix footer
