+++ app/components/Footer.tsx
@@ -6,19 +6,18 @@ import { FaInstagram } from "react-icons/fa"; // Gunakan react-icons
 
 const Footer = () => {
   return (
-    <footer className="bg-black text-white p-8">
+    <footer className="bg-black text-white p-8 mt-5">
       {/* Padding footer */}
-      <div className="flex flex-col items-center space-y-4">
+      <div className="flex flex-col justify center items-start">
         {/* Menambah space-y untuk memberi jarak antar elemen */}
         {/* Gunakan Image dari Next.js */}
         <Image
           src="/logo-1.png" // Path gambar logo
           alt="Airlangga Convention 2.0 Logo"
           width={150} // Tentukan lebar gambar
-          height={43} // Tentukan tinggi gambar
-          className="mb-4" // Menambah margin bawah logo agar tidak terlalu rapat
+          height={43} // Tentukan tinggi gambar // Menambah margin bawah logo agar tidak terlalu rapat
         />
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-12 text-center md:text-left px-4">
+        <div className="flex flex-row gap-4 items-center justify-between w-full mb-11">
           {/* About Us Section */}
           <div>
             <h2 className="font-bold mb-3">About Us</h2>{" "}
@@ -50,7 +49,8 @@ const Footer = () => {
           </div>
 
           {/* Contact Us Section */}
-          <div>
+        </div>
+        <div>
             <h2 className="font-bold mb-3">Contact Us</h2>{" "}
             {/* Menambah margin bawah pada judul */}
             <p className="flex items-center justify-center space-x-2">
@@ -58,12 +58,11 @@ const Footer = () => {
               <span>Instagram</span>
             </p>
           </div>
-        </div>
         {/* Divider */}
         <div className="border-t border-white w-full mt-6"></div>{" "}
         {/* Jarak lebih kecil antara divider dan konten */}
         {/* Copyright / Footer Text */}
-        <p className="mt-4 text-center">Aircon2.0</p>{" "}
+        <p className="mt-4 text-center">&copy;Aircon2.0</p>{" "}
         {/* Jarak yang lebih besar antara divider dan teks */}
       </div>
     </footer>
